% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dagger-package.R
\docType{package}
\name{dagger-package}
\alias{dagger}
\alias{dagger-package}
\title{dagger: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://asshah4.github.io/dagger}
  \item \url{https://github.com/asshah4/dagger}
  \item Report bugs at \url{https://github.com/asshah4/dagger}
}

}
\author{
\strong{Maintainer}: Anish Sanjay Shah \email{ashah282@uic.edu} (\href{https://orcid.org/0000-0002-9729-1558}{ORCID})

}
\keyword{internal}
