% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list-of-formulas.R
\name{formula_helpers}
\alias{formula_helpers}
\alias{get_left_vars}
\alias{get_left_side}
\alias{get_right_vars}
\alias{get_right_side}
\title{Tools for working with formulas}
\usage{
get_left_vars(x)

get_left_side(x)

get_right_vars(x)

get_right_side(x)
}
\description{
Tools for working with formulas
}
