% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/links.R
\name{links}
\alias{links}
\alias{link}
\alias{link.character}
\alias{link.tm}
\alias{link.default}
\title{Link together two terms}
\usage{
link(x = unspecified(), ...)

\method{link}{character}(x, to, ...)

\method{link}{tm}(x, to, ...)

\method{link}{default}(x = unspecified(), ...)
}
\description{
Link together two terms
}
