% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list-of-formulas.R
\name{list_of_formulas}
\alias{list_of_formulas}
\title{List of formulas}
\usage{
list_of_formulas(...)
}
\description{
Vectorized subclass of the \code{list_of} class from {vctrs::list_of()}
}
\section{Pluralized Arguments}{
For the arguments that would be dispatched for objects that are plural,
e.g. containing multiple terms such as a \code{formula} object, the input should
be wrapped within a \code{list()}.

For example, for the \strong{role} argument, it would be written:

\code{role = list(X ~ "exposure", M ~ "mediator", C ~ "confounder")}

This applies for all others plural objects and arguments.
}

