% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paths.R
\name{paths}
\alias{paths}
\alias{paths.character}
\alias{paths.default}
\title{paths vector}
\usage{
paths(x = unspecified(), ...)

\method{paths}{character}(
  x = character(),
  y = character(),
  role_x = character(),
  role_y = character(),
  label_x = character(),
  label_y = character(),
  ...
)

\method{paths}{default}(x = unspecified(), ...)
}
\description{
paths vector
}
