% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survival.R
\name{tbl_categorical_hazard}
\alias{tbl_categorical_hazard}
\title{Make complex regression table}
\usage{
tbl_categorical_hazard(
  object,
  data,
  events = formula(),
  followup = character(),
  terms = formula(),
  adjustment = formula(),
  ...
)
}
\description{
Make complex regression table
}
