% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gt.R
\name{tbl_forest}
\alias{tbl_forest}
\title{Subgroup Forest Plot}
\usage{
tbl_forest(object, ...)
}
\arguments{
\item{x}{Class \code{forge} object with models that have been fit}

\item{formula}{Identifies the relationships of interest, with LHS
representing the outcome, and RHS representing the exposure.}

\item{vars}{Character vector of the variables of interest. They can either be
grouping variables, or can be performed as interaction terms (based on the
models included originally). If the parameter \strong{interaction} is changed to
\code{TRUE}, then will create joint confidence intervals for the presence or
absence of the interaction term.}

\item{interaction}{Logical value to identify if the model should attempt to
use interaction terms or not. Currently only accepts binary/dichotomous
variables as the interaction term.}

\item{columns}{Additional columns that help to describe the subgroup models.
At least one column should be selected from this list. The sequence listed #'   will reflect the sequence shown in the table. The current options are:
\itemize{
\item beta = point estimate value, such as odds ratio or hazard ratio
\item conf = inclusion of the confidence interval (presumed to be 95\%-ile)
\item n = number of observations in each model group
}

For example: \verb{list(beta ~ "Hazard", conf ~ "95\% CI" n ~ "No.")"}}

\item{flip}{Determine if the odds or hazard ratio should be shown as the
reciprocal values. Instead of a decreasing hazard for every unit increase,
it describes an increasing hazard for every unit decrease.}

\item{axis}{Argument to help modify the forest plot itself. This is a
list-formula of the following parameters. If they are not named, the
function will attempt to "guess" the optimal parameters. The options are:
\itemize{
\item title = label or title for the column describing the forest plot
\item lim = x-axis limits
\item breaks = x-axis tick marks or break points that should be numbers
\item int = x-axis intercept
\item lab = label for the x-axis
\item scale = defaults to continuous, but may also use a log transformation as
well \code{c("continuous", "log")}
}

For example: \verb{list(title ~ "Decreasing Hazard", lab ~ "HR (95\% CI))}}

\item{width}{Describes the width of each column in a list of two-sided
formulas. The RHS is a decimal reflecting the percent each column should
take of the entire table. The forest plot is usually given 30\% of the
width.

For example: \code{list(n ~ .1, forest ~ 0.3)}}
}
\description{
Subgroup Forest Plot
}
