% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gt.R
\name{theme_gt_compact}
\alias{theme_gt_compact}
\title{Compact and minimal theme for \code{gt} tables}
\usage{
theme_gt_compact(
  data,
  table.font.size = gt::pct(80),
  table.width = gt::pct(90),
  ...
)
}
\arguments{
\item{data}{A table object that is created using the \code{\link[gt:gt]{gt()}} function.}

\item{table.width}{The width of the table. Can be specified as a
single-length character with units of pixels or as a percentage. If
provided as a single-length numeric vector, it is assumed that the value is
given in units of pixels. The \code{\link[gt:px]{px()}} and \code{\link[gt:pct]{pct()}} helper functions can also
be used to pass in numeric values and obtain values as pixel or percent
units.}

\item{...}{For passing additional arguments to the \code{\link[gt:tab_options]{gt::tab_options()}}
function}
}
\description{
This theme was used for placing somewhat larger tables into \code{xaringan} slides
by making the spacing more compact and decreasing the font size. The exposed
variables are to control font size and table width, but any option from the
\code{gt} package is allowed.
}
\concept{visualizers}
