% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/durandal-package.R
\docType{package}
\name{durandal-package}
\alias{durandal}
\alias{durandal-package}
\title{durandal: Causal Chains and Connections}
\description{
...
}
\seealso{
Useful links:
\itemize{
  \item \url{https://asshah4.github.io/durandal}
  \item Report bugs at \url{https://github.com/asshah4/durandal/issues}
}

}
\author{
\strong{Maintainer}: Anish Sanjay Shah \email{ashah282@uic.edu} (\href{https://orcid.org/0000-0002-9729-1558}{ORCID})

}
\keyword{internal}
