% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list-of-formulas.R
\name{formula_helpers}
\alias{formula_helpers}
\alias{lhs}
\alias{rhs}
\alias{rhs.formula}
\alias{lhs.formula}
\title{Tools for working with formula-like objects}
\usage{
lhs(x, ...)

rhs(x, ...)

\method{rhs}{formula}(x, ...)

\method{lhs}{formula}(x, ...)
}
\description{
Tools for working with formula-like objects
}
