% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{path_stack}
\alias{path_stack}
\alias{stack_paths}
\alias{stack_paths.vctrs_path}
\title{Path stack vector}
\usage{
stack_paths(f, ...)

\method{stack_paths}{vctrs_path}(f, roles, labels, ...)
}
\description{
Path stack vector
}
