% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paths.R
\name{paths}
\alias{paths}
\alias{paths.character}
\alias{paths.default}
\title{Pathways}
\usage{
paths(x = unspecified(), ...)

\method{paths}{character}(
  x = character(),
  to = character(),
  role = list(),
  label = list(),
  family = character(),
  ...
)

\method{paths}{default}(x = unspecified(), ...)
}
\arguments{
\item{x}{An object of the following types}

\item{role}{Specific roles the variable plays within the formula. These are
of particular importance, as they serve as special terms that can effect
how a formula is interpreted. The options for roles are as below:
\itemize{
\item \strong{outcome}: outcome/dependent variable that serves as an individual
variable in the \eqn{exposure -> outcome} relationship (DEFAULT for left
sided variables)
\item \strong{predictor}: predictor variable that is non-specific but decidedly on
the right-hand side of an equation (DEFAULT for right sided variables)
\item \strong{confounder}: predictor variable that is thought to be a confounder of
the causal relationship in the \eqn{exposure <- confounder -> outcome}
pathway, normally thought of as an adjustment or controlling variable
\item \strong{exposure}: predictor variable that serves as a primary or key
variable in the \eqn{exposure -> outcome} relationship
\item \strong{mediator}: predictor variable that is thought to be a causal
interm_archetypeediary in the \eqn{exposure -> mediator -> outcome} pathway
\item \strong{unknown}: default role of a variable that has not yet been assigned a
place, such as a potential intermediary object
}}

\item{label}{Display-quality label describing the variable}
}
\description{
Pathways
}
\section{List~Formula Arguments}{


The \strong{role} and \strong{label} arguments should be provided as a list of formulas.
}

