% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paths.R
\name{paths}
\alias{paths}
\alias{paths.character}
\alias{paths.formula}
\alias{paths.script}
\alias{paths.default}
\title{Pathways}
\usage{
paths(x = unspecified(), ...)

\method{paths}{character}(x, to = character(), role = list(), tier = list(), label = list(), ...)

\method{paths}{formula}(x, role = list(), tier = list(), label = list(), ...)

\method{paths}{script}(x, ...)

\method{paths}{default}(x = unspecified(), ...)
}
\arguments{
\item{x}{An object of the following types:
\itemize{
\item \code{character} where \strong{x} is the start of the \emph{path} and \strong{to} is the end
of the \emph{path}
\item \code{formula} from \code{\link[stats:formula]{stats::formula()}}
\item \code{script} object from \code{\link[archetypes:script]{archetypes::prescribe()}}
}}

\item{role}{Specific roles the variable plays within the formula. These are
of particular importance, as they serve as special terms that can effect
how a formula is interpreted. The options for roles are as below:
\itemize{
\item \strong{outcome}: outcome/dependent variable that serves as an individual
variable in the \eqn{exposure -> outcome} relationship (DEFAULT for left
sided variables)
\item \strong{predictor}: predictor variable that is non-specific but decidedly on
the right-hand side of an equation (DEFAULT for right sided variables)
\item \strong{confounder}: predictor variable that is thought to be a confounder of
the causal relationship in the \eqn{exposure <- confounder -> outcome}
pathway, normally thought of as an adjustment or controlling variable
\item \strong{exposure}: predictor variable that serves as a primary or key
variable in the \eqn{exposure -> outcome} relationship
\item \strong{mediator}: predictor variable that is thought to be a causal
interm_archetypeediary in the \eqn{exposure -> mediator -> outcome} pathway
\item \strong{unknown}: default role of a variable that has not yet been assigned a
place, such as a potential intermediary object
}}

\item{tier}{Grouping variable name for independent variables for modeling
terms together}

\item{label}{Display-quality label describing the variable}
}
\description{
Pathways
}
\section{List~Formula Arguments}{


The \strong{role} and \strong{label} arguments should be provided as a list of
formulas. For the arguments that would be dispatched for objects that are
plural, e.g. containing multiple terms such as a \code{formula} object, the input
should be wrapped within a \code{list()}.

For example, for the \strong{role} argument, it would be written:

\code{role = list(X ~ "exposure", Y ~ "outcome", M ~ "mediator")}

This applies for all others plural objects and arguments.
}

