% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gt.R
\name{tbl_group_forests}
\alias{tbl_group_forests}
\title{Grouped Forest Plots}
\usage{
tbl_group_forests(
  object,
  formula,
  type,
  vars,
  level = list(),
  columns = list(beta ~ "Estimate", conf ~ "95\% CI", n ~ "No."),
  flip = FALSE,
  axis = list(scale ~ "continuous"),
  width = list()
)
}
\arguments{
\item{object}{A \code{forge} object that contains or represents model data, with
required grouping variables (such as strata or interaction terms).}

\item{formula}{Identifies the relationships of interest, with LHS
representing the outcome, and RHS representing the exposure.}

\item{type}{Character vector to identify if the model was using subgroups or
interaction terms. The difference is that confidence intervals for
interaction can be calculated in multiple ways, and the effect size is
dependent on the presence or absence of the interaction variable.
\itemize{
\item \strong{interaction} = assumes an interaction term that is binary/dichotomous
\item \strong{subgroup} = assumes a categorical variable that was used to group the
original dataset
}}

\item{vars}{Character vector of the variables of interest. They can either be
grouping variables, or can be performed as interaction terms (based on the
models included originally). If the parameter \strong{interaction} is changed to
\code{TRUE}, then will create joint confidence intervals for the presence or
absence of the interaction term.}

\item{level}{List of formulas. Each list-element is a formula with the LHS
reflecting either the variable to re-label or a specific level, and the RHS
reflecting what the new level should be called (for display). If there are
conflicting labels, the most recent will be used.

For example, \verb{list(am ~ c("Manual", "Automatic")} would take, from the
\code{mtcars} dataset, the \code{am} variable, which consists of \code{c(0, 1)}, and
relabel them in the order described. They are sorted in ascendiing order
prior to re-labeling.

The alternative approach is to use the specific level itself and have it
re-labeled. \code{list(0 ~ "Absent")} would take all levels that are zero, and
change their value.}

\item{columns}{Additional columns that help to describe the subgroup models.
At least one column should be selected from this list. The sequence listed
will reflect the sequence shown in the table. The current options are:
\itemize{
\item beta = point estimate value, such as odds ratio or hazard ratio
\item conf = inclusion of the confidence interval (presumed to be 95\%-ile)
\item n = number of observations in each model group
\item p = p.value for model or interaction term
}

For example: \verb{list(beta ~ "Hazard", conf ~ "95\% CI" n ~ "No.")"}}

\item{flip}{Determine if the odds or hazard ratio should be shown as the
reciprocal values. Instead of a decreasing hazard for every unit increase,
it describes an increasing hazard for every unit decrease.}

\item{axis}{Argument to help modify the forest plot itself. This is a
list-formula of the following parameters. If they are not named, the
function will attempt to "guess" the optimal parameters. The options are:
\itemize{
\item title = label or title for the column describing the forest plot
\item lim = x-axis limits
\item breaks = x-axis tick marks or break points that should be numbers
\item int = x-axis intercept
\item lab = label for the x-axis
\item scale = defaults to continuous, but may also use a log transformation as
well \code{c("continuous", "log")}
}

For example: \verb{list(title ~ "Decreasing Hazard", lab ~ "HR (95\% CI))}}

\item{width}{Describes the width of each column in a list of two-sided
formulas. The RHS is a decimal reflecting the percent each column should
take of the entire table. The forest plot is usually given 30\% of the
width.

For example: \code{list(n ~ .1, forest ~ 0.3)}}
}
\description{
Grouped Forest Plots
}
